@@ -18,8 +18,9 @@
         "source": "https://github.com/ali-translator/buffered-translation"
     },
     "require": {
-        "php": ">=7.1",
+        "php": "^7.4",
         "ali-translator/translator":  "^0",
+        "ali-translator/text-template": "^0",
         "ext-intl": "*"
     },
     "autoload": {
@@ -34,7 +35,7 @@
     },
     "require-dev": {
         "symfony/var-dumper": "^5.0",
-        "php": ">=7.3.0",
+        "php": "^7.4",
         "phpunit/phpunit": "^9.0"
     },
     "suggest": {
